Module google_discovery1::api

Structs

ApiGetRestCall
Retrieve the description of a particular version of an api.

ApiListCall
Retrieve the list of APIs supported at this endpoint.

ApiMethods
A builder providing access to all methods supported on api resources. It is not used directly, but through the Discovery hub.

DirectoryList
There is no detailed description.

DirectoryListItems
The individual directory entries. One entry per api/version pair.

DirectoryListItemsIcons
Links to 16x16 and 32x32 icons representing the API.

Discovery
Central instance to access all Discovery related resource activities

JsonSchema
There is no detailed description.

JsonSchemaAnnotations
Additional information about this property.

JsonSchemaVariant
In a variant data type, the value of one property is used to determine how to interpret the entire entity. Its value must exist in a map of descriminant values to schema names.

JsonSchemaVariantMap
The map of discriminant value to schema to use for parsing..

RestDescription
There is no detailed description.

RestDescriptionAuth
Authentication information.

RestDescriptionAuthOauth2
OAuth 2.0 authentication information.

RestDescriptionAuthOauth2Scopes
The scope value.

RestDescriptionIcons
Links to 16x16 and 32x32 icons representing the API.

RestMethod
There is no detailed description.

RestMethodMediaUpload
Media upload parameters.

RestMethodMediaUploadProtocols
Supported upload protocols.

RestMethodMediaUploadProtocolsResumable
Supports the Resumable Media Upload protocol.

RestMethodMediaUploadProtocolsSimple
Supports uploading as a single HTTP request.

RestMethodRequest
The schema for the request.

RestMethodResponse
The schema for the response.

RestResource
There is no detailed description.