Struct google_discovery1::RestMethod

pub struct RestMethod {
    pub scopes: Option<Vec<String>>,
    pub description: Option<String>,
    pub parameters: Option<HashMap<String, JsonSchema>>,
    pub supports_media_upload: Option<bool>,
    pub etag_required: Option<bool>,
    pub media_upload: Option<RestMethodMediaUpload>,
    pub request: Option<RestMethodRequest>,
    pub use_media_download_service: Option<bool>,
    pub http_method: Option<String>,
    pub supports_subscription: Option<bool>,
    pub parameter_order: Option<Vec<String>>,
    pub id: Option<String>,
    pub path: Option<String>,
    pub response: Option<RestMethodResponse>,
    pub supports_media_download: Option<bool>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
scopes: Option<Vec<String>>

OAuth 2.0 scopes applicable to this method.

description: Option<String>

Description of this method.

parameters: Option<HashMap<String, JsonSchema>>

Details for all parameters in this method.

supports_media_upload: Option<bool>

Whether this method supports media uploads.

etag_required: Option<bool>

Whether this method requires an ETag to be specified. The ETag is sent as an HTTP If-Match or If-None-Match header.

media_upload: Option<RestMethodMediaUpload>

Media upload parameters.

request: Option<RestMethodRequest>

The schema for the request.

use_media_download_service: Option<bool>

Indicates that downloads from this method should use the download service URL (i.e. "/download"). Only applies if the method supports media download.

http_method: Option<String>

HTTP method used by this method.

supports_subscription: Option<bool>

Whether this method supports subscriptions.

parameter_order: Option<Vec<String>>

Ordered list of required parameters, serves as a hint to clients on how to structure their method signatures. The array is ordered such that the "most-significant" parameter appears first.

id: Option<String>

A unique ID for this method. This property can be used to match methods between different versions of Discovery.

path: Option<String>

The URI path of this REST method. Should be used in conjunction with the basePath property at the api-level.

response: Option<RestMethodResponse>

The schema for the response.

supports_media_download: Option<bool>

Whether this method supports media downloads.