Struct google_discovery1::JsonSchema

pub struct JsonSchema {
    pub description: Option<String>,
    pub format: Option<String>,
    pub enum_: Option<Vec<String>>,
    pub variant: Option<JsonSchemaVariant>,
    pub enum_descriptions: Option<Vec<String>>,
    pub read_only: Option<bool>,
    pub minimum: Option<String>,
    pub repeated: Option<bool>,
    pub id: Option<String>,
    pub ref_: Option<String>,
    pub default: Option<String>,
    pub items: Option<Option<Box<JsonSchema>>>,
    pub required: Option<bool>,
    pub maximum: Option<String>,
    pub properties: Option<HashMap<String, JsonSchema>>,
    pub location: Option<String>,
    pub pattern: Option<String>,
    pub additional_properties: Option<Option<Box<JsonSchema>>>,
    pub type_: Option<String>,
    pub annotations: Option<JsonSchemaAnnotations>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

description

A description of this object.

format

An additional regular expression or key that helps constrain the value. For more details see: http://tools.ietf.org/html/draft-zyp-json-schema-03#section-5.23

enum_

Values this parameter may take (if it is an enum).

variant

In a variant data type, the value of one property is used to determine how to interpret the entire entity. Its value must exist in a map of descriminant values to schema names.

enum_descriptions

The descriptions for the enums. Each position maps to the corresponding value in the "enum" array.

read_only

The value is read-only, generated by the service. The value cannot be modified by the client. If the value is included in a POST, PUT, or PATCH request, it is ignored by the service.

minimum

The minimum value of this parameter.

repeated

Whether this parameter may appear multiple times.

id

Unique identifier for this schema.

ref_

A reference to another schema. The value of this property is the "id" of another schema.

default

The default value of this property (if one exists).

items

If this is a schema for an array, this property is the schema for each element in the array.

required

Whether the parameter is required.

maximum

The maximum value of this parameter.

properties

If this is a schema for an object, list the schema for each property of this object.

location

Whether this parameter goes in the query or the path for REST requests.

pattern

The regular expression this parameter must conform to. Uses Java 6 regex format: http://docs.oracle.com/javase/6/docs/api/java/util/regex/Pattern.html

additional_properties

If this is a schema for an object, this property is the schema for any additional properties with dynamic keys on this object.

type_

The value type for this schema. A list of values can be found here: http://tools.ietf.org/html/draft-zyp-json-schema-03#section-5.1

annotations

Additional information about this property.