Struct google_discovery1::RestDescription

pub struct RestDescription {
    pub protocol: Option<String>,
    pub methods: Option<HashMap<String, RestMethod>>,
    pub labels: Option<Vec<String>>,
    pub kind: Option<String>,
    pub canonical_name: Option<String>,
    pub documentation_link: Option<String>,
    pub owner_name: Option<String>,
    pub package_path: Option<String>,
    pub batch_path: Option<String>,
    pub id: Option<String>,
    pub features: Option<Vec<String>>,
    pub owner_domain: Option<String>,
    pub root_url: Option<String>,
    pub name: Option<String>,
    pub parameters: Option<HashMap<String, JsonSchema>>,
    pub icons: Option<RestDescriptionIcons>,
    pub version_module: Option<bool>,
    pub description: Option<String>,
    pub title: Option<String>,
    pub exponential_backoff_default: Option<bool>,
    pub base_url: Option<String>,
    pub etag: Option<String>,
    pub version: Option<String>,
    pub service_path: Option<String>,
    pub discovery_version: Option<String>,
    pub schemas: Option<HashMap<String, JsonSchema>>,
    pub auth: Option<RestDescriptionAuth>,
    pub base_path: Option<String>,
    pub resources: Option<HashMap<String, RestResource>>,
    pub revision: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get rest apis (response)

Fields

protocol

The protocol described by this document.

methods

API-level methods for this API.

labels

Labels for the status of this API, such as labs or deprecated.

kind

The kind for this response.

canonical_name

Indicates how the API name should be capitalized and split into various parts. Useful for generating pretty class names.

documentation_link

A link to human readable documentation for the API.

owner_name

The name of the owner of this API. See ownerDomain.

package_path

The package of the owner of this API. See ownerDomain.

batch_path

The path for REST batch requests.

id

The ID of this API.

features

A list of supported features for this API.

owner_domain

The domain of the owner of this API. Together with the ownerName and a packagePath values, this can be used to generate a library for this API which would have a unique fully qualified name.

root_url

The root URL under which all API services live.

name

The name of this API.

parameters

Common parameters that apply across all apis.

icons

Links to 16x16 and 32x32 icons representing the API.

version_module

no description provided

description

The description of this API.

title

The title of this API.

exponential_backoff_default

Enable exponential backoff for suitable methods in the generated clients.

base_url

[DEPRECATED] The base URL for REST requests.

etag

The ETag for this response.

version

The version of this API.

service_path

The base path for all REST requests.

discovery_version

Indicate the version of the Discovery API used to generate this doc.

schemas

The schemas for this API.

auth

Authentication information.

base_path

[DEPRECATED] The base path for REST requests.

resources

The resources in this API.

revision

The version of this API.